requirements.txt
# requirements.txt

# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=9.5.0

# Math Engine
sympy>=1.12

# ML Utilities
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web App
streamlit>=1.28.0

# Data
pandas>=2.0.0
tqdm>=4.65.0

# Optional - for advanced features
# transformers>=4.30.0   # For using pretrained Vision Transformers
# onnxruntime>=1.15.0    # For model optimization